test(io): move admin server tests to test_admin.py
Towards #4946.
